Refactored some lazy methods code into a separte file
