+++ src/java_bytecode/java_bytecode_language.cpp
@@ -29,6 +29,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "java_bytecode_parser.h"
 #include "java_class_loader.h"
 #include "java_utils.h"
+#include "gather_methods_lazily.h"
 
 #include "expr2java.h"
 
@@ -170,269 +171,6 @@ bool java_bytecode_languaget::parse(
   return false;
 }
 
-/// Find a virtual callee, if one is defined and the callee type is known to
-/// exist.
-/// \par parameters: `needed_classes`: set of classes that can be instantiated.
-///   Any potential callee not in this set will be ignored.
-/// `call_basename`: unqualified function name with type signature (e.g.
-///   "f:(I)")
-/// `classname`: class name that may define or override a function named
-///   `call_basename`.
-/// `symbol_table`: global symtab
-/// \return Returns the fully qualified name of `classname`'s definition of
-///   `call_basename` if found and `classname` is present in `needed_classes`,
-///   or irep_idt() otherwise.
-static irep_idt get_virtual_method_target(
-  const std::set<irep_idt> &needed_classes,
-  const irep_idt &call_basename,
-  const irep_idt &classname,
-  const symbol_tablet &symbol_table)
-{
-  // Program-wide, is this class ever instantiated?
-  if(!needed_classes.count(classname))
-    return irep_idt();
-  auto methodid=id2string(classname)+"."+id2string(call_basename);
-  if(symbol_table.has_symbol(methodid))
-    return methodid;
-  else
-    return irep_idt();
-}
-
-/// Find possible callees, excluding types that are not known to be
-/// instantiated.
-/// \par parameters: `c`: function call whose potential target functions should
-///   be determined.
-/// `needed_classes`: set of classes that can be instantiated. Any potential
-///   callee not in this set will be ignored.
-/// `symbol_table`: global symtab
-/// `class_hierarchy`: global class hierarchy
-/// \return Populates `needed_methods` with all possible `c` callees, taking
-///   `needed_classes` into account (virtual function overrides defined on
-///   classes that are not 'needed' are ignored)
-static void get_virtual_method_targets(
-  const code_function_callt &c,
-  const std::set<irep_idt> &needed_classes,
-  std::vector<irep_idt> &needed_methods,
-  symbol_tablet &symbol_table,
-  const class_hierarchyt &class_hierarchy)
-{
-  const auto &called_function=c.function();
-  assert(called_function.id()==ID_virtual_function);
-
-  const auto &call_class=called_function.get(ID_C_class);
-  assert(!call_class.empty());
-  const auto &call_basename=called_function.get(ID_component_name);
-  assert(!call_basename.empty());
-
-  auto old_size=needed_methods.size();
-
-  auto child_classes=class_hierarchy.get_children_trans(call_class);
-  for(const auto &child_class : child_classes)
-  {
-    auto child_method=
-      get_virtual_method_target(
-        needed_classes,
-        call_basename,
-        child_class,
-        symbol_table);
-    if(!child_method.empty())
-      needed_methods.push_back(child_method);
-  }
-
-  irep_idt parent_class_id=call_class;
-  while(1)
-  {
-    auto parent_method=
-      get_virtual_method_target(
-        needed_classes,
-        call_basename,
-        parent_class_id,
-        symbol_table);
-    if(!parent_method.empty())
-    {
-      needed_methods.push_back(parent_method);
-      break;
-    }
-    else
-    {
-      auto findit=class_hierarchy.class_map.find(parent_class_id);
-      if(findit==class_hierarchy.class_map.end())
-        break;
-      else
-      {
-        const auto &entry=findit->second;
-        if(entry.parents.empty())
-          break;
-        else
-          parent_class_id=entry.parents[0];
-      }
-    }
-  }
-
-  if(needed_methods.size()==old_size)
-  {
-    // Didn't find any candidate callee. Generate a stub.
-    std::string stubname=id2string(call_class)+"."+id2string(call_basename);
-    symbolt symbol;
-    symbol.name=stubname;
-    symbol.base_name=call_basename;
-    symbol.type=c.function().type();
-    symbol.value.make_nil();
-    symbol.mode=ID_java;
-    symbol_table.add(symbol);
-  }
-}
-
-/// See output
-/// \par parameters: `e`: expression tree to search
-/// \return Populates `result` with pointers to each function call within e that
-///   calls a virtual function.
-static void gather_virtual_callsites(
-  const exprt &e,
-  std::vector<const code_function_callt *> &result)
-{
-  if(e.id()!=ID_code)
-    return;
-  const codet &c=to_code(e);
-  if(c.get_statement()==ID_function_call &&
-     to_code_function_call(c).function().id()==ID_virtual_function)
-    result.push_back(&to_code_function_call(c));
-  else
-    forall_operands(it, e)
-      gather_virtual_callsites(*it, result);
-}
-
-/// See output
-/// \par parameters: `e`: expression tree to search
-/// `symbol_table`: global symtab
-/// \return Populates `needed` with global variable symbols referenced from `e`
-///   or its children.
-static void gather_needed_globals(
-  const exprt &e,
-  const symbol_tablet &symbol_table,
-  symbol_tablet &needed)
-{
-  if(e.id()==ID_symbol)
-  {
-    // If the symbol isn't in the symbol table at all, then it is defined
-    // on an opaque type (i.e. we don't have the class definition at this point)
-    // and will be created during the typecheck phase.
-    // We don't mark it as 'needed' as it doesn't exist yet to keep.
-    auto findit=symbol_table.symbols.find(to_symbol_expr(e).get_identifier());
-    if(findit!=symbol_table.symbols.end() &&
-       findit->second.is_static_lifetime)
-    {
-      needed.add(findit->second);
-    }
-  }
-  else
-    forall_operands(opit, e)
-      gather_needed_globals(*opit, symbol_table, needed);
-}
-
-static void initialize_needed_classes_from_pointer(
-  const pointer_typet &pointer_type,
-  const namespacet &ns,
-  const class_hierarchyt &ch,
-  ci_lazy_methodst &lazy_methods);
-
-/// See output
-/// \par parameters: `class_type`: root of class tree to search
-/// `ns`: global namespace
-/// \return Populates `lazy_methods` with all Java reference types reachable
-///   starting at `class_type`. For example if `class_type` is
-///   `symbol_typet("java::A")` and A has a B field, then `B` (but not `A`) will
-///   noted as a needed class.
-static void gather_field_types(
-  const typet &class_type,
-  const namespacet &ns,
-  const class_hierarchyt &ch,
-  ci_lazy_methodst &lazy_methods)
-{
-  const auto &underlying_type=to_struct_type(ns.follow(class_type));
-  for(const auto &field : underlying_type.components())
-  {
-    if(field.type().id()==ID_struct || field.type().id()==ID_symbol)
-      gather_field_types(field.type(), ns, ch, lazy_methods);
-    else if(field.type().id()==ID_pointer)
-    {
-      // Skip array primitive pointers, for example:
-      if(field.type().subtype().id()!=ID_symbol)
-        continue;
-      initialize_needed_classes_from_pointer(
-        to_pointer_type(field.type()), ns, ch, lazy_methods);
-    }
-  }
-}
-
-/// Build up list of methods for types for a specific pointer type. See
-/// `initialize_needed_classes` for more details.
-/// \param pointer_type: The type to gather methods for.
-/// \param ns: global namespace
-/// \param ch: global class hierarchy
-/// \param [out] lazy_methods: Populated with all Java reference types whose
-///   references may be passed, directly or indirectly, to any of the functions
-///   in `entry_points
-static void initialize_needed_classes_from_pointer(
-  const pointer_typet &pointer_type,
-  const namespacet &ns,
-  const class_hierarchyt &ch,
-  ci_lazy_methodst &lazy_methods)
-{
-  const symbol_typet &class_type=to_symbol_type(pointer_type.subtype());
-  const auto &param_classid=class_type.get_identifier();
-  std::vector<irep_idt> class_and_parents=
-    ch.get_parents_trans(param_classid);
-
-  class_and_parents.push_back(param_classid);
-
-  for(const auto &classid : class_and_parents)
-    lazy_methods.add_needed_class(classid);
-
-  gather_field_types(
-    pointer_type.subtype(), ns, ch, lazy_methods);
-}
-
-/// Build up a list of methods whose type may be passed around reachable
-/// from the entry point.
-/// \param entry_points: list of fully-qualified function names that
-///   we should assume are reachable
-/// \param ns: global namespace
-/// \param ch: global class hierarchy
-/// \param pointer_type_selector: The pointer_type_selector used to find out
-///   what other classes may be used for pointers
-/// \param [out] lazy_methods: Populated with all Java reference types whose
-///   references may be passed, directly or indirectly, to any of the functions
-///   in `entry_points`.
-static void initialize_needed_classes(
-  const std::vector<irep_idt> &entry_points,
-  const namespacet &ns,
-  const class_hierarchyt &ch,
-  ci_lazy_methodst &lazy_methods)
-{
-  for(const auto &mname : entry_points)
-  {
-    const auto &symbol=ns.lookup(mname);
-    const auto &mtype=to_code_type(symbol.type);
-    for(const auto &param : mtype.parameters())
-    {
-      if(param.type().id()==ID_pointer)
-      {
-        const pointer_typet &original_pointer=to_pointer_type(param.type());
-        initialize_needed_classes_from_pointer(
-          original_pointer, ns, ch, lazy_methods);
-      }
-    }
-  }
-
-  // Also add classes whose instances are magically
-  // created by the JVM and so won't be spotted by
-  // looking for constructors and calls as usual:
-  lazy_methods.add_needed_class("java::java.lang.String");
-  lazy_methods.add_needed_class("java::java.lang.Class");
-  lazy_methods.add_needed_class("java::java.lang.Object");
-}
-
 bool java_bytecode_languaget::typecheck(
   symbol_tablet &symbol_table,
   const std::string &module)
@@ -500,54 +238,6 @@ bool java_bytecode_languaget::typecheck(
   return false;
 }
 
-/// Translates the given list of method names from human-readable to
-/// internal syntax.
-/// Expands any wildcards (entries ending in '.*') in the given method
-/// list to include all non-static methods defined on the given class.
-/// \param [in, out] methods: List of methods to expand. Any wildcard entries
-///   will be deleted and the expanded entries appended to the end.
-/// \param symbol_table: global symbol table
-static void resolve_method_names(
-  std::vector<irep_idt> &methods,
-  const symbol_tablet &symbol_table)
-{
-  std::vector<irep_idt> new_methods;
-  for(const irep_idt &method : methods)
-  {
-    const std::string &method_str=id2string(method);
-    if(!has_suffix(method_str, ".*"))
-    {
-      std::string error_message;
-      irep_idt internal_name=
-        resolve_friendly_method_name(
-          method_str,
-          symbol_table,
-          error_message);
-      if(internal_name==irep_idt())
-        throw "entry point "+error_message;
-      new_methods.push_back(internal_name);
-    }
-    else
-    {
-      irep_idt classname="java::"+method_str.substr(0, method_str.length()-2);
-      if(!symbol_table.has_symbol(classname))
-        throw "wildcard entry point '"+method_str+"': unknown class";
-
-      for(const auto &name_symbol : symbol_table.symbols)
-      {
-        if(name_symbol.second.type.id()!=ID_code)
-          continue;
-        if(!to_code_type(name_symbol.second.type).has_this())
-          continue;
-        if(has_prefix(id2string(name_symbol.first), id2string(classname)))
-          new_methods.push_back(name_symbol.first);
-      }
-    }
-  }
-
-  methods=std::move(new_methods);
-}
-
 /// Uses a simple context-insensitive ('ci') analysis to determine which methods
 /// may be reachable from the main entry point. In brief, static methods are
 /// reachable if we find a callsite in another reachable site, while virtual
@@ -565,155 +255,30 @@ bool java_bytecode_languaget::do_ci_lazy_method_conversion(
   symbol_tablet &symbol_table,
   lazy_methodst &lazy_methods)
 {
-  class_hierarchyt ch;
-  ch(symbol_table);
-
-  std::vector<irep_idt> method_worklist1;
-  std::vector<irep_idt> method_worklist2;
-
-  auto main_function=
-    get_main_symbol(symbol_table, main_class, get_message_handler(), true);
-  if(main_function.stop_convert)
+  const auto method_converter=[&](
+    const symbolt &symbol,
+    const java_bytecode_parse_treet::methodt &method,
+    ci_lazy_methodst new_lazy_methods)
   {
-    // Failed, mark all functions in the given main class(es) reachable.
-    std::vector<irep_idt> reachable_classes;
-    if(!main_class.empty())
-      reachable_classes.push_back(main_class);
-    else
-      reachable_classes=main_jar_classes;
-    for(const auto &classname : reachable_classes)
-    {
-      const auto &methods=
-        java_class_loader.class_map.at(classname).parsed_class.methods;
-      for(const auto &method : methods)
-      {
-        const irep_idt methodid="java::"+id2string(classname)+"."+
-          id2string(method.name)+":"+
-          id2string(method.signature);
-        method_worklist2.push_back(methodid);
-      }
-    }
-  }
-  else
-    method_worklist2.push_back(main_function.main_function.name);
+    java_bytecode_convert_method(
+      symbol,
+      method,
+      symbol_table,
+      get_message_handler(),
+      max_user_array_length,
+      safe_pointer<ci_lazy_methodst>::create_non_null(&new_lazy_methods),
+      string_preprocess);
+  };
 
-  // Add any extra entry points specified; we should elaborate these in the
-  // same way as the main function.
-  std::vector<irep_idt> extra_entry_points=lazy_methods_extra_entry_points;
-  resolve_method_names(extra_entry_points, symbol_table);
-  method_worklist2.insert(
-    method_worklist2.begin(),
-    extra_entry_points.begin(),
-    extra_entry_points.end());
+  gather_methods_lazilyt method_gather(
+    symbol_table,
+    main_class,
+    main_jar_classes,
+    lazy_methods_extra_entry_points,
+    java_class_loader,
+    get_message_handler());
 
-  std::set<irep_idt> needed_classes;
-
-  {
-    std::vector<irep_idt> needed_clinits;
-    ci_lazy_methodst initial_lazy_methods(
-      needed_clinits,
-      needed_classes,
-      symbol_table);
-    initialize_needed_classes(
-      method_worklist2,
-      namespacet(symbol_table),
-      ch,
-      initial_lazy_methods);
-    method_worklist2.insert(
-      method_worklist2.end(),
-      needed_clinits.begin(),
-      needed_clinits.end());
-  }
-
-  std::set<irep_idt> methods_already_populated;
-  std::vector<const code_function_callt *> virtual_callsites;
-
-  bool any_new_methods;
-  do
-  {
-    any_new_methods=false;
-    while(!method_worklist2.empty())
-    {
-      std::swap(method_worklist1, method_worklist2);
-      for(const auto &mname : method_worklist1)
-      {
-        if(!methods_already_populated.insert(mname).second)
-          continue;
-        auto findit=lazy_methods.find(mname);
-        if(findit==lazy_methods.end())
-        {
-          debug() << "Skip " << mname << eom;
-          continue;
-        }
-        debug() << "CI lazy methods: elaborate " << mname << eom;
-        const auto &parsed_method=findit->second;
-        // Note this wraps *references* to method_worklist2, needed_classes:
-        ci_lazy_methodst lazy_methods(
-          method_worklist2,
-          needed_classes,
-          symbol_table);
-        java_bytecode_convert_method(
-          *parsed_method.first,
-          *parsed_method.second,
-          symbol_table,
-          get_message_handler(),
-          max_user_array_length,
-          safe_pointer<ci_lazy_methodst>::create_non_null(&lazy_methods),
-          string_preprocess);
-        gather_virtual_callsites(
-          symbol_table.lookup(mname).value,
-          virtual_callsites);
-        any_new_methods=true;
-      }
-      method_worklist1.clear();
-    }
-
-    // Given the object types we now know may be created, populate more
-    // possible virtual function call targets:
-
-    debug() << "CI lazy methods: add virtual method targets ("
-            << virtual_callsites.size()
-            << " callsites)"
-            << eom;
-
-    for(const auto &callsite : virtual_callsites)
-    {
-      // This will also create a stub if a virtual callsite has no targets.
-      get_virtual_method_targets(
-        *callsite,
-        needed_classes,
-        method_worklist2,
-        symbol_table,
-        ch);
-    }
-  }
-  while(any_new_methods);
-
-  // Remove symbols for methods that were declared but never used:
-  symbol_tablet keep_symbols;
-
-  for(const auto &sym : symbol_table.symbols)
-  {
-    if(sym.second.is_static_lifetime)
-      continue;
-    if(lazy_methods.count(sym.first) &&
-       !methods_already_populated.count(sym.first))
-    {
-      continue;
-    }
-    if(sym.second.type.id()==ID_code)
-      gather_needed_globals(sym.second.value, symbol_table, keep_symbols);
-    keep_symbols.add(sym.second);
-  }
-
-  debug() << "CI lazy methods: removed "
-          << symbol_table.symbols.size() - keep_symbols.symbols.size()
-          << " unreachable methods and globals"
-          << eom;
-
-  symbol_table.swap(keep_symbols);
-
-  return false;
+  return method_gather(symbol_table, lazy_methods, method_converter);
 }
 
 /// Provide feedback to `language_filest` so that when asked for a lazy method,
